hitbox: refactor
